+++ packages/BilibiliRecorder/src/index.ts
@@ -17,6 +17,8 @@ import type {
   RecorderCreateOpts,
   RecorderProvider,
   RecordHandle,
+  VideoFileCreatedPayload,
+  FormatName,
 } from "@bililive-tools/manager";
 
 function createRecorder(opts: RecorderCreateOpts): Recorder {
@@ -59,12 +61,16 @@ function createRecorder(opts: RecorderCreateOpts): Recorder {
         ...info,
       };
     },
-    async getStream() {
+    async getStream(opts: { formatName?: FormatName } = {}) {
       const res = await getStream({
         channelId: this.channelId,
         quality: this.quality,
-        formatName: this.formatName,
+        formatName: opts.formatName || this.formatName,
         codecName: this.codecName,
+        cookie: this.auth,
+        strictQuality: false,
+        onlyAudio: this.onlyAudio,
+        customHost: this.customHost,
       });
       return res.currentStream;
     },
@@ -273,7 +279,12 @@ const checkLiveStatusAndRecord: Recorder["checkLiveStatusAndRecord"] = async fun
     },
   );
 
-  const handleVideoCreated = async ({ filename, title, cover, rawFilename }) => {
+  const handleVideoCreated = async ({
+    filename,
+    title,
+    cover,
+    rawFilename,
+  }: VideoFileCreatedPayload) => {
     this.emit("videoFileCreated", { filename, cover, rawFilename });
 
     if (title && this?.liveInfo) {