+++ packages/DouYinRecorder/src/index.ts
@@ -187,6 +187,11 @@ function createRecorder(opts: RecorderCreateOpts): Recorder {
         quality: this.quality,
         streamPriorities: this.streamPriorities,
         sourcePriorities: this.sourcePriorities,
+        strictQuality: false,
+        formatPriorities: this.formatPriorities,
+        doubleScreen: this.doubleScreen,
+        api: this.api as APIType,
+        uid: this.uid,
       });
       return res.currentStream;
     },