+++ packages/app/src/main/index.ts
@@ -61,6 +61,9 @@ contextMenu({
   showSearchWithGoogle: false,
   showSaveImageAs: false,
 });
+if (process.argv.includes("--disable-gpu")) {
+  app.disableHardwareAcceleration();
+}
 
 const WindowState = new Store<{
   winBounds: {
@@ -222,7 +225,11 @@ function createWindow(): void {
     } else {
       // 用来静默启动
       const isHidden = process.argv.includes("--hidden");
-      if (!isHidden) mainWindow.show();
+      if (isHidden) {
+        mainWin.setSkipTaskbar(true);
+      } else {
+        mainWindow.show();
+      }
     }
   });
 
@@ -365,6 +372,7 @@ function createWindow(): void {
     {
       label: "显示",
       click: () => {
+        mainWin.setSkipTaskbar(false);
         mainWin.show();
       },
     },