+++ packages/app/src/renderer/src/pages/Tools/pages/Recorder/Index.vue
@@ -1,6 +1,9 @@
 <template>
   <div class="container">
-    <div style="display: flex; align-items: center; gap: 10px; margin-bottom: 20px">
+    <div
+      style="display: flex; align-items: center; gap: 10px; margin-bottom: 20px"
+      class="filter-container"
+    >
       <n-input
         v-model:value="params.name"
         placeholder="备注或房间号"
@@ -15,7 +18,7 @@
         clearable
       />
       <n-select
-        v-model:value="params.recordStatus"
+        v-model:value="params.status"
         :options="statusOptions"
         placeholder="录制状态"
         style="width: 140px"
@@ -98,11 +101,11 @@
             <div class="section" @click="edit(item.id)">直播间设置</div>
             <div class="section" @click="refresh(item.id)">刷新直播间信息</div>
             <div
-              v-if="item.recordHandle?.savePath && false"
+              v-if="item.living"
               class="section"
-              @click="open(item.id, item?.recordHandle?.url)"
+              @click="open(item.id, item.owner || item.remarks)"
             >
-              打开直播
+              观看直播
             </div>
             <div
               v-if="!isWeb"
@@ -151,7 +154,6 @@
       v-model:visible="batchOperateModalVisible"
       @completed="handleBatchOperateCompleted"
     ></batchOperateModal>
-    <videoModal :id="editId" v-model:visible="videoModalVisible" :video-url="videoUrl"></videoModal>
   </div>
 </template>
 
@@ -163,19 +165,19 @@ import addModal from "./components/addModal.vue";
 import batchAddModal from "./components/batchAddModal.vue";
 import batchResultModal from "./components/batchResultModal.vue";
 import batchOperateModal from "./components/batchOperateModal.vue";
-import videoModal from "./components/videoModal.vue";
 import cardView from "./components/cardView.vue";
 import listView from "./components/listView.vue";
 import { useRouter } from "vue-router";
 import ButtonGroup from "@renderer/components/ButtonGroup.vue";
 import ColumnSelector from "@renderer/components/ColumnSelector.vue";
 import { platformOptions } from "./data";
+import SortButton from "./components/SortButton.vue";
 
 import { useEventListener, useStorage } from "@vueuse/core";
 import eventBus from "@renderer/utils/eventBus";
+import { toLiveVideoPlayerPage } from "@renderer/utils/pages";
 
 import type { RecorderAPI } from "@biliLive-tools/http/types/recorder.js";
-import SortButton from "./components/SortButton.vue";
 
 defineOptions({
   name: "recorder",
@@ -228,7 +230,7 @@ const recorderLocalParams = useStorage(
 
 const params = ref<Parameters<typeof recoderApi.infoList>[0]>({
   platform: undefined,
-  recordStatus: undefined,
+  status: undefined,
   name: undefined,
   autoCheck: undefined,
   page: 1,
@@ -240,8 +242,16 @@ const statusOptions = ref([
     value: "recording",
   },
   {
-    label: "未录制",
-    value: "unrecorded",
+    label: "空闲中",
+    value: "idle",
+  },
+  {
+    label: "检查错误",
+    value: "check-error",
+  },
+  {
+    label: "标题被屏蔽",
+    value: "title-blocked",
   },
 ]);
 const recordOptions = ref([
@@ -500,22 +510,23 @@ const edit = async (id: string) => {
   addModalVisible.value = true;
 };
 
-const videoModalVisible = ref(false);
-const videoUrl = ref("");
 /**
  * 打开直播间
  * @param id 内部直播间id
+ * @param owner 直播间主人名
  */
-const open = async (id: string, streamUrl: string) => {
-  editId.value = id;
-  videoUrl.value = streamUrl;
-  if (!streamUrl) {
-    notice.error({
-      title: "未找到直播流地址",
+const open = async (id: string, owner: string) => {
+  const info = await refresh(id, false);
+  if (info?.living === false) {
+    notice.warning({
+      title: "直播间未开播",
     });
     return;
   }
-  videoModalVisible.value = true;
+  toLiveVideoPlayerPage({
+    liveId: id,
+    owner: owner,
+  });
 };
 
 const getLiveInfo = async (forceRequest: boolean = false) => {
@@ -561,7 +572,7 @@ const getLiveInfo = async (forceRequest: boolean = false) => {
 };
 
 // 刷新单个直播间信息
-const refresh = async (id: string) => {
+const refresh = async (id: string, showNotification: boolean = true) => {
   const recorder = recorderList.value.find((item) => item.id === id);
   if (!recorder) return;
 
@@ -576,10 +587,12 @@ const refresh = async (id: string) => {
       [refreshedLiveInfo],
     );
   }
-
-  notice.success({
-    title: "刷新成功",
-  });
+  if (showNotification) {
+    notice.success({
+      title: "刷新成功",
+    });
+  }
+  return refreshedLiveInfo;
 };
 
 const handleModalClose = () => {
@@ -802,6 +815,11 @@ const handleActionClick = (key?: string | number) => {
     &.section-danger {
       color: var(--color-danger-text);
     }
+  }
+}
+@media (max-width: 1024px) {
+  .filter-container {
+    flex-wrap: wrap;
   }
 }